requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.2
plotly==5.18.0
python-jose==3.3.0
httpx==0.25.2
python-multipart==0.0.6 